Restore original dice component
